Split classes into 3 files. Updated namespace from `Silvershop` to `SilverShop`. Added branch alias. Added psr-4 autoloader directive.
