+++ README.md
@@ -8,7 +8,7 @@ Allows you to create a CMS button for creating and editing a single related obje
 
 In Warehouse.php context:
 ```php
-    use Silvershop\HasOneField\HasOneButtonField;
+    use SilverShop\HasOneField\HasOneButtonField;
 
 	public function getCMSFields() {
 		$fields = parent::getCMSFields();
