+++ src/HasOneButtonField.php
@@ -1,16 +1,14 @@
 <?php
 
-namespace Silvershop\HasOneField;
+namespace SilverShop\HasOneField;
 
-use SilverStripe\Control\Controller;
 use SilverStripe\Forms\GridField\GridField;
-use SilverStripe\Forms\GridField\GridFieldAddNewButton;
 use SilverStripe\Forms\GridField\GridFieldConfig;
 use SilverStripe\Forms\GridField\GridFieldDetailForm;
-use SilverStripe\Forms\GridField\GridField_HTMLProvider;
-use SilverStripe\ORM\DataList;
-use SilverStripe\View\ArrayData;
 
+/**
+ * Class HasOneButtonField
+ */
 class HasOneButtonField extends GridField
 {
 
@@ -31,56 +29,5 @@ class HasOneButtonField extends GridField
     public function getRecord()
     {
         return $this->record;
-    }
-}
-
-class GridFieldHasOneEditButton extends GridFieldAddNewButton implements GridField_HTMLProvider
-{
-
-    public function getHTMLFragments($gridField)
-    {
-        $record = $gridField->getRecord();
-        if (!$record->exists() || !$record->isInDB()) {
-            return parent::getHTMLFragments($gridField); //use parent add button
-        }
-        $singleton = singleton($gridField->getModelClass());
-        if (!$singleton->canCreate()) {
-            return array();
-        }
-        if (!$this->buttonName) {
-            // provide a default button name, can be changed by calling {@link setButtonName()} on this component
-            $objectName = $singleton->i18n_singular_name();
-            $this->buttonName = _t('GridField.Edit', 'Edit {name}', array('name' => $objectName));
-        }
-        $data = new ArrayData(array(
-            'NewLink' => Controller::join_links($gridField->Link('item'), $record->ID, 'edit'),
-            'ButtonName' => $this->buttonName,
-        ));
-
-        return array(
-            $this->targetFragment => $data->renderWith('SilverStripe\\Forms\\GridField\\GridFieldAddNewButton')
-        );
-    }
-}
-
-class HasOneButtonRelationList extends DataList
-{
-
-    protected $record;
-    protected $name;
-    protected $parent;
-
-    public function __construct($record, $name, $parent)
-    {
-        $this->record = $record;
-        $this->name = $name;
-        $this->parent = $parent;
-        parent::__construct($record->ClassName);
-    }
-
-    public function add($item)
-    {
-        $this->parent->{$this->name."ID"} = $item->ID;
-        $this->parent->write();
     }
 }
